[Feat] #197 - interceptor가 정상적으로 작동하도록 수정
getUserType에서 unregistered 반환하도록 수정
