+++ SOPT-iOS/Projects/Data/Sources/Repository/SignInRepository.swift
@@ -38,7 +38,9 @@ extension SignInRepository: SignInRepositoryInterface {
                 else {
                     return self.userService.reissuance()
                 }
-                
+                // NOTE: (@준호) 플그 미등록 + 비활동 유저의 경우 임시로 accessToken 빈 스트링 부여
+                // 자동로그인 시 활용하기 위함
+                UserDefaultKeyList.Auth.appAccessToken = ""
                 return Fail(error: error).eraseToAnyPublisher()
             })
             .map { entity in
