--- SOPT-iOS/Projects/Modules/Network/Sources/Foundation/AFManager.swift
@@ -1,23 +1,0 @@
-//
-//  AFManager.swift
-//  Network
-//
-//  Created by Junho Lee on 2022/10/16.
-//  Copyright © 2022 SOPT-Stamp-iOS. All rights reserved.
-//
-
-import Foundation
-
-import Alamofire
-
-public class DefaultAlamofireManager: Alamofire.Session {
-    static let shared: DefaultAlamofireManager = {
-        let configuration = URLSessionConfiguration.default
-        configuration.timeoutIntervalForRequest = 10
-        configuration.timeoutIntervalForResource = 10
-        configuration.requestCachePolicy = .reloadIgnoringLocalCacheData
-        let interceptor = AlamoInterceptor()
-        let session = DefaultAlamofireManager(configuration: configuration, interceptor: interceptor)
-        return session
-    }()
-}
